Added Much Better Color Boundry Numbers

+++ src/main/java/frc/robot/wheel/SenseColor.java
@@ -25,7 +25,7 @@ public class SenseColor extends SubsystemBase {
   private final I2C.Port i2cPort = I2C.Port.kOnboard;
   private final ColorSensorV3 m_colorSensor = new ColorSensorV3(i2cPort);
   private int proximity = m_colorSensor.getProximity();
-  private int IR = m_colorSensor.getIR();
+  private double IR = m_colorSensor.getIR();
   private Color detectedColor = m_colorSensor.getColor();
   private final ColorMatch m_colorMatcher = new ColorMatch();
 
@@ -41,7 +41,7 @@ public class SenseColor extends SubsystemBase {
     return detectedColor;
   }
 
-  public int getRawColor() {
+  public double getRawColor() {
     return IR;
   }
